bot/requirements.txt
python-telegram-bot==20.7
google-cloud-secret-manager==2.16.4
google-auth==2.23.4
google-api-python-client==2.108.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0